+++ config.js
@@ -1,6 +1,6 @@
 window.CONTACTFLOW_CONFIG = {
-  edition: 'Personal Ultimate 3.4',
-  version: '3.4.0',
+  edition: 'Personal Ultimate 3.5',
+  version: '3.5.0',
   googleClientId: '',
   telegramNativeAppConfigured: false
 };